Split user types
