+++ packages/types/src/discord/guilds.ts
@@ -1,12 +1,13 @@
 /** Types for: https://discord.com/developers/docs/resources/guild */
 
-import type { DiscordAvatarDecorationData, DiscordRole, DiscordUser, DiscordVoiceState, OAuth2Scope } from '../discord.js'
+import type { DiscordRole, DiscordVoiceState, OAuth2Scope } from '../discord.js'
 import type { DiscordChannel } from './channels.js'
 import type { DiscordEmoji } from './emojis.js'
 import type { DiscordPresenceUpdate } from './gateway.js'
 import type { DiscordSoundboardSound } from './soundboards.js'
 import type { DiscordStageInstance } from './stageInstances.js'
 import type { DiscordSticker } from './stickers.js'
+import type { DiscordAvatarDecorationData, DiscordUser } from './users.js'
 
 /** https://discord.com/developers/docs/resources/guild#guild-object-guild-structure */
 export interface DiscordGuild {
