+++ packages/types/src/discord/messages.ts
@@ -1,6 +1,5 @@
 /** Types for: https://discord.com/developers/docs/resources/message */
 
-import type { DiscordUser } from '../discord.js'
 import type { DiscordApplication } from './applications.js'
 import type { DiscordChannel, DiscordThreadMember } from './channels.js'
 import type { DiscordEmoji } from './emojis.js'
@@ -8,6 +7,7 @@ import type { DiscordMember } from './guilds.js'
 import type { DiscordAuthorizingIntegrationOwners, DiscordMessageComponents, DiscordMessageInteraction, InteractionTypes } from './interactions.js'
 import type { DiscordPoll } from './polls.js'
 import type { DiscordSticker, DiscordStickerItem } from './stickers.js'
+import type { DiscordUser } from './users.js'
 
 /** https://discord.com/developers/docs/resources/message#message-object-message-structure */
 export interface DiscordMessage {
