+++ packages/types/src/index.ts
@@ -19,6 +19,7 @@ export * from './discord/soundboards.js'
 export * from './discord/stageInstances.js'
 export * from './discord/stickers.js'
 export * from './discord/subscriptions.js'
+export * from './discord/users.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
 export * from './shared.js'
