+++ packages/types/src/shared.ts
@@ -2,33 +2,6 @@ import type { Locales } from './discord/reference.js'
 
 export type BigString = bigint | string
 
-/** https://discord.com/developers/docs/resources/user#user-object-premium-types */
-export enum PremiumTypes {
-  None,
-  NitroClassic,
-  Nitro,
-  NitroBasic,
-}
-
-/** https://discord.com/developers/docs/resources/user#user-object-user-flags */
-export enum UserFlags {
-  DiscordEmployee = 1 << 0,
-  PartneredServerOwner = 1 << 1,
-  HypeSquadEventsMember = 1 << 2,
-  BugHunterLevel1 = 1 << 3,
-  HouseBravery = 1 << 6,
-  HouseBrilliance = 1 << 7,
-  HouseBalance = 1 << 8,
-  EarlySupporter = 1 << 9,
-  TeamUser = 1 << 10,
-  BugHunterLevel2 = 1 << 14,
-  VerifiedBot = 1 << 16,
-  EarlyVerifiedBotDeveloper = 1 << 17,
-  DiscordCertifiedModerator = 1 << 18,
-  BotHttpInteractions = 1 << 19,
-  ActiveDeveloper = 1 << 22,
-}
-
 /** https://discord.com/developers/docs/topics/permissions#role-object-role-flags */
 export enum RoleFlags {
   None,
